[6/10] change ExposeEvent to WindowEvent
The 'ExposeEvent' is sent when (part of) a window was previously hidden
and now becomes visible (=exposed).

In SDL2 this event is part of the more general SDL_WindowEvent. Later
patches in this series add support for the more general window-event
(because ImGui needs it). As a preparation for that step, this patch
already changes ExposeEvent to WindowEvent. IOW WindowEvent is only
partially supported yet (only the window-wxpose sub-event is), it will
be fully supported later.