[8/10] Switch keyboard events to SDL_Event
This patch converts our keyboards events to use SDL_Event internally.
Similarly to what was done before for joystick and mouse events.

This patch is the largest in the series because simultaneously with
converting the event, we needed to change how key-presses are
represented:
* SDL2 has 'SDL_Keycode' and 'SDL_Scancode' these represent a single key
  without modifiers. It also has 'SDL_Keysym' which is the combination
  of a keycode and a scancode plus modifiers. SDL2 offers functions to
  convert SDL_Keycode and SDL_Scancode to and from a string, but there
  are no such functions for 'SDL_Keysym'.
* In openMSX we had a 'Keycode' class which is more or less the
  combination of a SDL_Keycode plus modifiers plus press/release status
  (and we also abused it for scancodes). We had to/from string functions
  that do include the modifiers and press/release status.

For the MSX keyboard code (for the character mode) we need the
combination of unicode and keycode. That combination was present in
SDL1, but SDL2 splits that information in two different event types:
key-events and text-events. For the vast majority of the applications
that's a better approach, unfortunately not for openMSX. In openMSX we
have some heuristic that tries to reconstruct a SDL1-like key event
(thus including unicode info) by combining key- and text-events.

This patch introduces a new class 'SDLKey' as a replacement for the old
'KeyCode' class. SDLKey is the combination of a SDL_Keysym (so
key+scancode plus modifiers) plus a boolean to indicate press or
release. In addition we abuse an unused field in the SDL_Keysym struct
(a field with the name 'unsued') to store (reconstructed) unicode
information.

Together with this new 'SDLKey' class this patch introduces fromString()
and toString() functions that work in a backwards-compatible way (I mean
with the previous to/from string functions in openMSX). See comments in
the code itself for a lot more details on these conversion functions.

A large part of the patch is reworking the tables in the MSX keyboard
emulation code. Before the numerical values for the (old) 'Keycode'
were all relatively small/compact, meaning looking on keycode could be
done efficiently with a plain array. The numerical values for SDL_Keysym
and SDL_Scancode span a _much_ wider range (and the values are sparse),
thus lookup via an array is no longer feasible. A side-effect (bonus) of
this refactoring is that the tables are now split for key- and
scan-codes.