Save/Restore TreeNode-open-status for ImGuiVdpRegs
Extended our im::TreeNode() wrapper with functionality to force
open/close the TreeNode.

Use this to remember the open/collapsed status for the groups in the VDP
register viewer.

Showing all VDP registers takes up a lot of vertical space. Sure you can
scroll through it, but usually you're only interested in specific
registers. The registers are subdivided in groups, and you can collapse
the groups you're not interested in. When continuing you're
debug-session (e.g. the next day) it's useful if this status is
retained.