Make it easier to load/save stuff from/to "imgui.ini"
We added extra sections in the 'imgui.ini' file. E.g. to store whether
various windows are open, and should be reopened when openMSX is
restarted. And many other things.

Saving a variable is typically only 1 line of code. Loading a variable
typically requires 2-4 lines of code. That's not a lot. Then why this
patch to simplify this further?

The main problems with this "manual" approach are:
* Both the save and load part need to specify a name. And this (string)
  needs to be the same for both. Unfortunately it's easy to make a
  mistake (typo, copy/paste error, ...). In the new system this mistake
  is impossible.
* Depending on the type of the variable (bool, int, float, ...) the load
  and save code looks different. If a refactoring changes the type of a
  variable, without a corresponding change in the load/save code, there
  was a good chance that code continues to compile fine, but it would do
  the wrong thing at-runtime. The new system automatically picks the
  correct implementation.
So in summary: the new system makes it harder to make mistakes. And the
fact that it's shorter (easier) to use is a nice bonus.

The new load/save system only handles simple static variables. Complex
types or dynamic variables (e.g. the context of a std::map) continue
using the exact same code we have today. Because we don't try to handle
everything the new load/save system is relatively simple (~120 lines of
code). (In contrast to our serialization code which does need to handle
all complex cases).