Implemented view::take() + unittest
This is a (minimal) reimplementation of c++20 std::views::take.

This functionality is present in c++20 and we do already use c++20 in
openMSX. Then why reimplement it? Unfortunately clang only supports it
starting from version 16, and the github runners for macOS are still
using clang-14 :(

In the future we'll replace our version with the standard version.