Increase minimum refresh rate from 10fps to 25fps
In openMSX screen redrawing is triggered by the emulation: when the
emulated MSX finishes a frame (at 50 or 60Hz) we redraw the screen. This
includes the MSX screen itself, but also all other stuff like OSD
elements and the ImGui stuff.

When the MSX is paused for some reason (real pause or breakpoint), or
when MSX is emulated at a slower speed (via 'speed' setting or because
host is too slow (possibly in combination with 'minframeskip' setting)),
there are no, or not enough, MSX frame endings.

To mitigate this problem we also have a mechanism to ensure a minimal
(host) refresh rate. When we've drawn a frame we start a timer, when
that timer expires we do draw another (host) frame. When a new MSX frame
arrives before the timer expires we only restart the timer. In other
words: when running normally this timer does nothing. Only when there
aren't (any / frequent-enough) MSX frame we insert extra host redraws.

Before this patch the above timer was set to 100ms, corresponding with a
minimum refresh rate of 10fps. Manuel rightfully complained that this
makes the new ImGui debugger and console sluggish when e.g. emulation
was breaked. This patch decreases the value to 40ms, corresponding to
25fps. To me this 'feels' OK now, but if needed we can decrease it
further, as long as we stay _below_ the MSX refresh rate (with a little
margin).

This value (10fps) was chosen _many_ years ago. Nowadays host computers
are much faster, and keeping up with 25 fps shouldn't be a problem
anymore. Especially not now that we've made OpenGL rendering mandatory.