Rename description of joytap/ninjatap connectors
Before this patch all 4 joytap connectors shared the same description
"Joy Tap port". This patch changes that to "Joy Tap port <N>" with N =
1..4. Similar for ninjatap.

This change was needed after the previous patch which displays the
connector description (instead of the connector name) in the GUI.

The original idea was:
* All 4 ports are equivalent, so they can have the same description.
* And to distinguish the different instances they do have a unique name.
But this idea breaks with the above GUI change.

(Because of this change we now have to store the each of the
descriptions in its own 'std::string' instead of a single shared
'static_string_view', but that's only a minor thing).