Fix parsing of sjasm symbol files
Reported by David. The following line was not recognized:
    setscreen2: EQU 0x0000435F
The problem was that 'equ' was only checked lower-case. Fixed by doing a
case-insensitive check (+ extended unittest).

The 8-digit hex value was not a problem. As long as the numerical value
fits in 16 bit. (It already worked, but I included a check for this in
the unittests).