Fix ImGui disassembly view icw long symbols
Reported by David:
When a (long) symbol is used the mnemonic column in the disassmbly view
looks like:
    ld (debugenemy
instead of
    ld (debugenemy_previous_gravity_value),hl

Fixed. The dasm() function was originally used for the openmsx 'debug
disasm' command. Now it's reused in the ImGui disassembly view (and
generalized so that it can optionally print addresses as labels). Though
the formatting (= truncate mnemonic to 19 chars) was left in the dasm()
function. I moved it out and only do it for the 'debug disasm' command.