Fix ImGuiOsdIcons
They were unusable when starting from a clean 'imgui.ini' file.

By default (with a clean 'imgui.ini') place the icons at the bottom left
(to not interfere with the main menu bar).

Don't allow right-click in the icons when the icons cannot be moved. In
other words: disable all mouse-interaction with the icons via this
setting.

Also several tweaks in the icon configuration window. And add more help
text.

Don't have the console open when starting openMSX for the first time
(with a clean 'imgui.ini').