Fix blurry-font issue in ImGui
Fixed in Dear ImGui opengl integration code. I'm working to get this fix
upstream (or an equivalent better fix).

The bug only triggered when compiling with '-ffast-math'. This flag is
enabled in the 'opt' build.

More in detail: the problem was caused by a rounding error when setting
up the openGL viewport:
  int width = 1280.0 * 1.0;
  results in 1279 instead of 1280
The problem was that '1.0' was not exactly 1.0 but 0.999999940395.
And that value was the result of:
   int width = 1280;
   int w = 1280;
   float scale = (float)width / w;
And with '-ffast-math' this did not give exactly 1.0.