Simplify rendering code: setting of viewport
While debugging the ImGui blurry font issue, I was surprised by how
openMSX sets the openGL viewport. There are two scenarios:
  a) Either it's setup once (*1).
  b) Or it's set on every frame (*2).
(*1) With once I mean once for a whole lot of frames. It's still reset
     when e.g. scale_factor changes or when switching from/to fullscreen.
(*2) In case of b) we _also_ do the then useless setup of a). And that
     was confusing to me: I was debugging with a wrong assumption about
     where in the code we set the viewport.

The difference between scenario a) and b) is whether we require
'render-to-texture'. We need this when either:
  * horizontal_stretch != 320
  * glow != 0
  * display_deform != 0
  * ...
The default for horizontal_stretch=280, thus usually (almost always?) we
use the render-to-texture code path.

With render-to-texture we first render the MSX screen (including the
effects of the scale_algorithm) to a texture, and then draw this texture
to the actual framebuffer (possibly with some further effects like
horizontal stretch). Without render-to-texture we directly render the
MSX screen to the framebuffer.

So skipping render-to-texture, when not needed, is faster. But with
current graphics hardware (and even with +15 year old graphics hardware)
the difference is negligible. And then removing this (minor) optimization
does simplify the code.

And this simplification is a small step towards having a freely
resizable window (as opposed to the discrete scale_factor's we have
now). That will also require a more dynamic viewport calculation.