Made 'gl::context' independent from screen resolution
Before gl::context was constructed with 'width,height' parameters. This
was (only) needed to setup the model-view-projection matrix.

Now move this calculation to a new method setupMvpMatrix(). And we call
this method each frame. (Possibly tweak: call it on window resize).

This is a preparation to reuse the existing GL-context when the window
is resized. And this is needed to fix a bug in ImGui on changing
scale_factor. And longer term it's a preparation to make the main
openMSX window freely resizable.