Fix disassembly view on step-back
On each debug-break or step we scroll the disassembly view to the PC
(more in detail: we scroll so that the PC appears at 1/4 of the full
height). However this didn't always work when stepping backwards.
Especially when the step-back was not to the previous instruction in
program memory (e.g. when stepping back over a jump, call, return).

After a debug-break (and this includes forward steps) we set a flag,
indicating that we should scroll to the new PC location. We do the same
for step-back, but we wrongly set this flag before instead of after
executing the step-back. And then of course that scrolls to the wrong
location.

Easy to fix once the problem was known. (But figuring this out was not
obvious (to me)).