Optimize disassembly view
Before we calculated the length of all instructions in the 64kB Z80
address space. Now we try to limit this to the visible instructions, and
some before (just enough to be able to calculate instruction
boundaries).

See comment in the patch for more details.

Before this patch we called 'instructionLength()' tens of thousands of
times per frame (typical number: ~50000). Now, on average, that's
reduced to less than 10.