Add location indicator in ImGui hex editor (when window has lost focus)
As suggested by David:
    When you edit data in the memory viewer or enter an address directly
    it is clear which byte you're looking at. However, if you continue
    debugging your program the memory viewer looses focus, and you do
    not have any indication anymore of the byte you were monitoring. So
    simply have the background of this byte in a slightly different
    color will help tremendously in finding the byte you're interested
    in during a debug session.

The hex editor is a 3rd-party widget. We do have the source code of this
widget, but my preference is to avoid changing this code. Luckily it did
offer a 'HighlightFn' callback.

Using this callback and some ('public' but 'internal') variables in the
widget, I was able to always show the preview-highlight, even when the
'OptShowDataPreview' option is disabled. And including when the window
has lost focus.

Side-note, to enable the data-preview, right-click and check 'Show Data
Preview'. Then you can select the data-type, and this influences the
number of bytes in the highlight (1, 2, 4 or 8 bytes).