Revert (small) part of f2e820604
There was a bug. With:
* scale_algorithm=simple
* scanline=0
* blur=0
the MSX screen was not rendered correctly (the typical symptom was:
screen remains black, but other things could happen as well). Thanks to
David for reporting this problem

Commit f2e820604 made initialization of 'GLContext' independent from the
screen resolution. One of the consequences is that stuff like the
model-view-projection matrix must be recalculated when the screen
resolution changes. This matrix is used by the shaders (= our scale
algorithms). So instead of setting this matrix once when the shader is
constructed, we now re-set it on every frame in the GLScaler::setup()
method.

Before commit f2e820604 we called this setup() method before every
vertical MSX screen 'region'. Where 'region' means a switch from
vertical border to display area, or a screensplit which changes between
256 and 512 pixel wide mode. The reason for these regions is that
typically some details of the scaler implementation are different
between regions.

In commit f2e820604 I made the following reasoning:
* The only parameter to setup() is 'bool superimposed'. This is constant
  for the whole frame. This parameter switches between two variations of
  the shader (with superimposing included or not).
* After f2e820604 we also need to set the mvp-matrix for the shader. But
  this also remains constant for the whole frame.
* So instead of calling 'currScaler->setup()' for each region it should
  be sufficient to call it only once per frame. (Because 'currScaler'
  also remains constant during the whole frame).

Unfortunately this still caused a problem:
* In some cases the 'currScaler' (this more or less corresponds to the
  shader program for the current scale_algorithm) decides that it itself
  is not suited to handle a particular region, and instead it delegates
  to the 'fallback' scaler.
   - This can happen e.g. because a scaler can only handle 256-wide
     modes, and not 512-wide modes.
   - Or for example the simple-scaler, with scaline=0 and blur=0 decides
     that it's more efficient to use an even simpler scaler.
* The 'fallback' scaler simply does nearest neighbour interpolation. (So
  indeed the effect of the 'simple' scaler without scanlines and without
  blur).
* However when delegating to the fallback scaler, we never called the
  setup() method for the fallback scaler. And the the mvp-matrix was not
  correct, which then (among other possible symptoms) results in a black
  screen.

Fixed by reverting a small part of f2e820604. We now keep calling
GLScaler::setup() for each region (internally from
GLScaler::scaleImage(), including the fallback scaler) instead of only
once per frame (from PostProcessor).