Fix display of symbols (e.g. in disassembly view)
The 'lookupValueCache' used a std::string_view to refer to the actual
names in the 'symbols' vector. Though when re-sorting the symbols in
the GUI, the symbols change position, and along with the symbols also
the symbol-names. Thus the string-views became invalid.

Fixed by storing a std::string (thus effectively a copy) of the
symbol-names.

Thanks to David for reporting this bug.