Recognition of .ogv files (for laserdisc) was broken ...
... after commit 8e561ba11. That commit changed:
    memcmp(packet.packet, "fishead", 8)
to
    ranges::equal(std::span{packet.packet, 8}, std::string_view("fishead"))

In general using ranges::equal() is more type-safe than memcmp(). (And
we caught actual bugs via cleanups similar to this one). Though in the
above transformation we overlooked that the old code used an implicit
0-terminator in the "fishead" string. After the transformation we
wrongly compare range of length 8 with a range of only length 7, and
those always compare not-equal.