Fixed crash when using 2 V9990 devices at once
This can (unintentionally) happen when e.g. using the Boosted_MSX2_EN
machine, which has a Video9000 builtin, in combination with the '-ext
video9000' command line option.

The problem were these 2 callbacks (introduced in 96e43d900a):
  v9990_invalid_register_read_callback
  v9990_invalid_register_write_callback
With two V9990 instances, this now creates two (pairs of) settings with
the same name. In an assert build this immediately triggers an
assertion. (In a non-assert build it appears to work fine, but no
guarantees).

Fixed by changing the name to
  <device-name>_invalid_register_read_callback
  <device-name>_invalid_register_write_callback
That's not backwards compatible, but this feature is only used rarely.
This should be fine.