Allow to store an empty string for settings in settings.xml
For example you might want to do this for the
'invalid_psg_directions_callback' setting.

The problem was that our Settings class has the weird concept of a value
that should never be saved to settings.xml. The only reason this exists
is to avoid having "renderer=none" saved to your settings.xml file.
(Maybe this mechanism can be removed in the imgui branch?).

Before this patch there was always at least one value that couldn't be
saved, and by default it's the empty string. This patch changes the type
of this value from 'TclObject' to 'std::optional<TclObject>' so that it
becomes possible to really have _no_ non-save value.