Add the "socket_port_number" and "socket_selection_mode" settings. These settings are for Windows only.
"socket_port_number" changes the base number used to choose a TCP
port to listen from incoming connections from the default of 9938.

"socket_selection_mode" has two possible values, "random" (default,
choose a random port between base and base+64) and "fixed" (use
the base port as the actual port number).

The chosen port number is still increased by one up to 64 times
in case the port is already in use, regardless of the value of
socket_selection_mode.

When the value of socket_port_number or socket_selection_mode changes,
the socket is closed and re-opened with a new port number according
to the new values of the settings. This also updates the socket
file accordingly.

The documentation is also updated appropriately.