Show error message when a symbol file could not be loaded
Most likely this may happen for stored filenames in 'imgui.ini' which
have been deleted the next time openMSX is started.

In the per-file section of the symbol manager, files that could not be
loaded are shown in a red color. When expanding the corresponding
section, also the error message is shown. The 'reload' button is shown
right below the error message (this button is also there for correctly
loaded files).