Support extensions in 'Cartrige slot X' menu
This handles extensions in a specific cartrige slot (corresponding to
the 'extX' commands). I/O-only extensions can also be inserted via the
menu, but they don't show up in this specific cartridge slot. Adding an
option to remove such I/O-only extensions is left for a future patch.

This patch also makes the 'extX' commands more uniform with the other
media commands:
* Before this patch, only 'ext<x> <extension-name>' was supported.
* It was an error when the corresponding slot was already occupied.
* Now we also support the syntax 'ext<x> insert <extension-name>.
* When there already is something inserted in the slot (rom or
  extension), it will first be removed before inserting the given
  extension.