Refactor ImGuiMedia
Many functions always took a pair of parameters:
* The media name (e.g. 'diska')
* The corresponding 'MediaInfo' struct.

But the name and the MediaInfo object already live together as
key-value pairs in a map. This patch moves the name into the MediaInfo
struct, changes the map to a set, and then only passes a single
parameter to many functions.

This also reduced the number of times the media name needs to be
repeated (e.g. it removed 2 redundant "cassetteplayer" strings).