Use current-media-path for start-directory in Open-file-dialog
The open-file-dialog to insert media (disk, rom, cassette, ...) already
remembered the last location of an 'accepted' file (when you click Ok).
This was already separate per media-instance (diska, diskb, carta, ...).
And this location was already saved for the next openmsx session.

Additionally the file-dialog allows to set custom bookmark locations.
That way you can easily go to your favorite directories containing
disks, roms, ...

This patch makes the following addition (suggested by Manuel):
When there already is a disk or rom inserted, start from that location
the next time the open-file-dialog is opened. Before this patch we
mostly already had that behavior, except when the media is changed via
another way than the GUI. For example:
* Via the console.
* Via an external application.
* By loading a savestate or replay.