Partial revert of Manuel's setting description rename
Amount of frames to always skip  -> Minimum frame-skip   (was 'minframeskip')
Maximum amount of frames to skip -> Maximum frame-skip   (was 'maxframeskip')

I agree that the original names (e.g. 'minframeskip') were too cryptic
for a GUI. But Manuel's description is too long (IMHO). The idea is that
if you hover over the (compact) setting description you get a longer
explanation. Maybe today those explanations aren't ideal (yet). It are
the descriptions we give in the c++ code for the settings, for example
see: src/video/RenderSettings.cc, line 56

Instead of changing the compact ImGui description we should change the
setting description in the c++ code. That's also more widely useful.