Added extra info in 'machine info' window
We only showed the info (directly) taken from the machine-config <info>
tag (aka the meta-data). Now we add some extra stuff:
* The amount of RAM
* The amount of VRAM
* The VDP type
* The number of disk drives
* The number of cartridge slots
This info is obtained by querying the MSXMotherboard object. So we had
to actually instantiate the machine. That's more expensive than just
parsing the machine configuration file. Therefor we perform these extra
queries when we test the validity of the machine, that already required
(trying to) instantiate the machine.