ImGuiMedia: refactor cartridge and disk menus
Before there was a 'simple' and an 'advanced' menu/window to insert
disks and cartridges. This patch unifies both.

Also for disks it allows to insert 'dir-as-disk' or 'ramdisk', and for
cartridges it more clearly distinguishes between rom-images and
extensions.

This patch is relatively large. This is because we want to have distinct
history (recent used items) for e.g. disk images and dir-as-dsk. This
required a large internal refactoring.