ImGuiMedia: usability tweak (experiment)
Before this patch, when selecting a dir-as-dsk, it was required to first
click the dis-as-dsk radio button, and only then could you click the
dir-as-dsk browse button (or recent items). Now when you click the
browse button it will automatically also select the radio button.

Maybe this is not standard GUI behavior. But personally I find this much
more usable. Let's try it as an experiment.