ImGuiMedia: show end of path when it doesn't fit on screen
For example:
   /home/wouter/msx/roms/Ashiguine 3 (1987) (Micro Cabin) (J).rom.gz
was shown (clipped) as:
   /home/wouter/msx/roms/Ashiguine 3 (1987) (Micro C
now we instead show
   ...ms/Ashiguine 3 (1987) (Micro Cabin) (J).rom.gz
when the window is too narrow.

Rationale: the last part of a path contains the filename, often that'
more informative that the start of the path which contains the
directory.

For now this is only done for the filename in rom and disk images. Let
me know if there are more locations where this should be applied.
Though one limitation is that we can only (easily?) do this for plain
text, not for text in an edit field.