ImGuiMachine: make machine displayName unique, tweak data-structure
In the previous patches we introduced a 'displayName' for machines. This
was calculated as '<manufacturer> <code>' (when meta-data is present in
the machine config). Unfortunately this name is no longer guaranteed to
be unique. E.g. this happens when a user copy-pastes a machine config
file, possibly makes some changes, but doesn't update the meta-data. (I
had several such machine configs).

Fixed by detecting such duplicate names, and if present append the
name of the config file. That one is guaranteed to be unique.

Before the last few patches we lazily retrieved the machine meta-data.
We only needed it when the detailed machine description was requested.
Though now we already need it to calculate the displayName. So there's
not much benefit anymore to lazily get it. And if we don't need that we
can simplify the data-structure to store all the machine related info.
(We do still lazily check whether a machine config is working or not).