ImGuiConsole: fix rendering of wider than 256 pixel (colored) chunks
The console input-line rendering uses a hack:
* The ImGui::TextInput widget doesn't supported colored text.
* But we do want to have syntax highlighting.
* To work around this, we first let ImGui render the widget (this is
  needed for e.g. the cursor, the selection, etc). Then we overdraw
  (only) the text in the correct color.

Overdrawing text is done in different chunks. Each time the text changes
color, we issue a new drawList->AddText() render command. Between chunks
we need to advance the x-coordinate. Before this patch we used the
function ImGui::CalcTextSize() to calculate how much to advance.
Unfortunately this function returns the wrong result (off-by-one) for
chunks wider than 256 pixels (read the comments in
src/3rdparty/imgui/imgui.cc:5416). Because of this the text was not
exactly overdrawn (but shifted one pixel), making it appear as bold.

Fixed by assuming a fixed-width font (this is indeed the case), and then
manually calculating the chunk-width (without rounding error).