Warn instead of crash on invalid keybindings in settings.xml
Before this patch, when there's an invalid keybinding saved in
'settings.xml', openMSX stopped with an error like:
    Uncaught exception: Invalid keycode: BlaBla

Now we only give a warning on these. But openMSX will continue to start.

Such invalid bindings can occur:
* When manually editing the 'settings.xml' file.
* When switching from a newer to an older openMSX release. And in
  particular when switching from the current 'imgui' to the 'master'
  branch, and then also creating a new binding for a key which has been
  renamed(*1).

(*1) After patch d24d37d66acb1c4b we use SDL2 names for keycodes instead
our own names. For many keys the name is identical, but for example our
old name "PRINT" is now "PrintScreen". The old names ARE still
recognized. That is: a newer openMSX version recognizes both the old and
the new names, but not the other way around: an older openMSX version
will not recognize the new names.