ImGuiMedia: show <title> instead of filename for ROMs known in the database
I think I replaced all rom filenames with the title. But let me know if
I missed some.

There's one thing I'm not yet happy with:
In the "Media > Cartridge Slot X" window. In the section "Select new
cartridge > ROM image", there's an edit box where you can edit the
filename, or select a recent ROM from a drop-down menu, or click the
browse button to open a select-file dialog.

Because this is an edit box we must show the filename, but in the
drop-down menu with recent items we instead show ROM titles instead of
filenames. And then when you select a recent items, you suddenly see a
filename instead of a title.

We can solve this in two ways:
1) Don't allow to edit the filename, then we can show the title instead
   of the filename.
2) Show filenames in the drop-down menu.
Any feedback?