ImGuiMedia: show "no disk drives present" ...
... when opening the "Media" menu on a MSX machine that has no disk
drives. Before this patch the menu would simply not show anything in the
disk drive section.

And similar for cartridge slots and the cassette port.

Note, when there's only one disk drive or one cartridge slot present, we
do NOT show a message about a missing 2nd drive or slot.