ImGuiMedia: make play/stop/record buttons reflect the openMSX cassetteplayer status
The corresponding play/stop/record button is drawn in a permanently
pressed way. Toggle buttons aren't directly supported by 'Dear ImGui',
but we can relatively easy fake it by temporarily changing the 'normal'
button color to the color of an 'active' button.

This patch also adds a utility function to conditionally change a
style-color. We already had utility functions to unconditionally (and
temporarily) change the colors. We use this new utility to implement the
new button behavior, but also to simplify some existing code.