On drag-and-drop a disk image: show what's present in the drive
When you drag-and-drop a disk image on the main openMSX window, and when
more than one disk drive is available, you get a popup to select the
drive for where to insert the disk.

Before this patch that popup had a selection list with just "diska" and
"diskb" (for 2 drives). Now that list shows what is already inserted in
the drives. For example:
    A: MSXDOS2.20.DSK
    B: Empty

Only the filename of the inserted disk is shown, not the whole path.
And similarly this patch now only shows the filename (not the whole
path) for ROM images (that are not known in the software database) in
the drag-and-drop for ROM images popup.