Make sure OSD widget name doesn't contain dots...
Sound chips that are inside a mapper type (e.g. HamarajaNight) are created with
the name of the ROM file, which typically contains a dot. The dot is also used
as a hierarchy separator character in the OSD widget stack. The script didn't
take that into account.
SolutionL replace any dot with an underscore in the chip name when creating the
OSD widgets.