Better (=no) tooltip for mute checkbox in tape deck
Extend the Checkbox(BooleanSetting&) utility function with an optional
parameter that allows to customize the tooltip that will be shown when
hovering the checkbox.

Use this to suppress the tooltip for the "Mute tape audio" checkbox in
the "Media > Tape deck" window. Before this patch the tooltip talked
about muting an individual soundchannel, this was too generic. Now
there's no tooltip at all. I guess the desription without extra tooltip
is clear enough. But anyone feel free to add a tooltip anyway.