ImGuiMedia: added (optional) filter for extensions
Similar to the filter for machines.
TODO: try to reduce code duplication between both filters. (I'm pushing
already to get early feedback on the useability).

Maybe one unusual thing about the UI is:
* when the filter-tree-node is expanded, we show a list box for the
  extensions
* when the filter-tree-code is collapsed, we show a combo-box.
Not sure it's a good idea, but it keeps the extension section compact
when the filter isn't used.