ImGuiMedia: fix crash on re-inserting the same ROM
To reproduce:
* Open 'Media > Cartridge Slot A'
* Select (any) rom image.
* Press 'Apply'
* Press 'Ok'
* Then hover over the 'Media > Cartridge Slot A' or
  'Media > Extensions > Remove' menu items.
Before this patch, this would crash.

The trigger is that 'Apply' already inserts the select ROM, and then
'Ok' inserts the same ROM again. Normally this isn't a problem:
inserting the ROM that is already inserted has no effect. Except for one
detail ...

When inserting a ROM, we first check that everything will work fine
(e.g. that the ROM file can be read) before removing the currently
inserted ROM. Therefor we first create the new HardwareConfig object
before destroying the old one. So for a short time there are two
'HardwareConfig' objects in memory for the same ROM (one that is
inserted, and one that is about to be inserted). HardwareConfig objects
have a name, and that name needs to be unique, making the name unique is
handled internally. The name for a ROM extension is the filename of the
ROM. But there can't be two with the same name, so the 2nd one gets a
suffix " (1)". This name is not visible in the GUI, it's typically only
used internally (and on the Tcl command line), so this is still perfecty
fine.

The crash occured later we when we want to query extra information on
this ROM extension (e.g. when hovering some menu items). This code
*wrongly* assumed that HardwareConfig::getName() always returned the ROM
filename. Often that's true, but not when the same ROM is used twice
(e.g. as in the above scenario, or when the same ROM is really inserted
in two cartridge slots).

Fixed by adding a new method HardwareConfig::getRomFilename() that
always returns the filename, even in case of duplicates.