[4/7] Added MSXJoystick
Largly a copy/paste of KeyJoystick with this important difference:
* Rather than only responding to (a single) keyboard key, the (6)
  buttons of the new MSXJoystick can be bound to any (list of)
  BooleanInputs.

In this patch the bindings are still hardcoded. Later patches in this
series will make them configurable.

This patch adds two new pluggables:
    msxjoystick1  and   msxjoystick2
The idea is that these can replace:
    keyjoystick1  and  keyjoystick2
But also all these:
    joystick<N>

Ideally next we would remove the now obsolete pluggables. But I'm afraid
that we need to keep them (in some form) for backwards compatibility
with old savestates and replay files. Still needs more investigation.
That's not part of this patch series. (Maybe we can keep them, for now,
but deprecated them, and not show them as pluggables in the GUI?)