[5/7] Made msxjoystick<n> mapping configurable via setting
The name of the setting is 'msxjoystick<n>_config' the format of the
setting is similar (but not 100% the same) as the (now deprecated)
joystick<n>_config settings.

More details after the next patch (then you can configure it in the GUI
and observe what effect this has on the value of the setting).