Remove ImGuiMarkdown, 3rdparty/imgui_md, 3rdparty/md4c
The idea was to use the 'imgui_md' library to display the openMSX
manuals directly in the openMSX application. Unfortunately, for now(?),
this library isn't advanced enough for this (yet?).

But we don't loose much by not doing this. Instead we just show a link
to the manuals. And maybe it's also better to read manuals in your
favorite browser than in openMSX itself.

This patch removes:
* The imgui_md library (a markdown renderer widget for Dear ImGui)
* The md4c library (a markdown parser, used by imgui_md)
* The ImGuiMarkdown class (our integration of imgui_md).

ImGuiMarkdown was the only user of the fonts:
    vera13  veraBold13  veraBold16  veraItalic13  veraBoldItalic13
So now we don't load these anymore at startup.

I did not remove the actual .ttf font files. Because in the future we
may want to experiment with using a proportional font in the GUI, rather
than the builtin Dear ImGui mono spaced font.