package-darwin: Install docs to the app Resources folder.
After commit ade6773555 openMSX would crash immediately after
starting, because when it looks for the documentation in the
openMSX.app/Contents/Resources/doc location it can’t find it
and throws an exception.

This change installs the docs in the expected location, which
makes the openMSX app boot properly again, and also the links
in the Help menu work as expected.

The only downside is that there is now no longer a separate
Documentation folder in the dmg archive next to the app,
since we can only specify one INSTALL_DOC_DIR. However I
think that’s fine, it can be found in other ways.